requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-rest-passwordreset==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
et-xmlfile==1.1.0
fsspec==2024.3.1
idna==3.6
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
PyJWT==2.8.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
